test the driver-rabbitmq in chaosControl
