+++ chaos-common/src/main/java/io/openchaos/common/utils/KillProcessUtil.java
@@ -46,4 +46,17 @@ public class KillProcessUtil {
             log.info("No {} process running in node {}.", processName, node);
         }
     }
+
+    public static void forceKillInErl(String node, String processName) throws Exception {
+        log.info("Force kill node {} process {} .", node, processName);
+        String pidList = SshUtil.execCommandWithArgsReturnStr(node, String.format("ps ax | grep -i '%s' | grep -v grep | awk '{print $1}'", processName)).trim();
+        if (!pidList.isEmpty()) {
+            String[] pids = pidList.split("\n");
+            for (String pid : pids) {
+                SshUtil.execCommand(node, String.format("kill -9 %s", pid));
+            }
+        } else {
+            log.info("No {} process running in node {}.", processName, node);
+        }
+    }
 }
