+++ chaos-common/src/main/java/io/openchaos/common/utils/PauseProcessUtil.java
@@ -48,4 +48,18 @@ public class PauseProcessUtil {
         }
 
     }
+
+    public static void resumeInErl(String node, String processName) throws Exception {
+        log.info("Resume node {} process {} .", node, processName);
+        String pidList = SshUtil.execCommandWithArgsReturnStr(node, String.format("ps ax | grep -i '%s'| grep -v grep | awk '{print $1}'", processName)).trim();
+        if (!pidList.isEmpty()) {
+            String[] pids = pidList.split("\n");
+            for (String pid : pids) {
+                SshUtil.execCommand(node, String.format("kill -CONT %s", pid));
+            }
+        } else {
+            log.info("No {} process running in node {}.", processName, node);
+        }
+
+    }
 }
