+++ chaos-framework/pom.xml
@@ -84,6 +84,11 @@
             <artifactId>driver-etcd</artifactId>
         </dependency>
         <dependency>
+            <groupId>io.openmessaging</groupId>
+            <artifactId>driver-rabbitmq</artifactId>
+            <version>0.6.5-SNAPSHOT</version>
+        </dependency>
+        <dependency>
             <groupId>com.panayotis</groupId>
             <artifactId>javaplot</artifactId>
         </dependency>
