+++ driver-rabbitmq/src/main/java/io/openchaos/driver/rabbitmq/config/RabbitMQClientConfig.java
@@ -1,0 +1,6 @@
+package io.openchaos.driver.rabbitmq.config;
+
+public class RabbitMQClientConfig {
+    public String user;
+    public String password;
+}
