+++ driver-rabbitmq/src/main/java/io/openchaos/driver/rabbitmq/config/RabbitMQConfig.java
@@ -1,0 +1,12 @@
+package io.openchaos.driver.rabbitmq.config;
+
+public class RabbitMQConfig {
+    public String installDir;
+    public String rabbitmqVersion;
+    public String configureFilePath;
+
+    public RabbitMQConfig setInstallDir(String dir){
+        installDir = dir;
+        return this;
+    }
+}
