+++ driver-rabbitmq/src/main/java/io/openchaos/driver/rabbitmq/core/HaMode.java
@@ -1,0 +1,9 @@
+package io.openchaos.driver.rabbitmq.core;
+
+public  enum HaMode {
+    classic, quorum;
+
+    HaMode(){
+
+    }
+}
