+++ driver-rabbitmq/src/main/java/io/openchaos/driver/rabbitmq/core/HaParams.java
@@ -1,0 +1,6 @@
+package io.openchaos.driver.rabbitmq.core;
+
+public class HaParams {
+    String totalNodes;
+    String nodes;
+}
