--- driver-rabbitmq/src/main/java/io/openmessaging/driver/rabbitmq/config/RabbitMQBrokerConfig.java
@@ -1,7 +1,0 @@
-package io.openmessaging.driver.rabbitmq.config;
-
-public class RabbitMQBrokerConfig {
-    public String brokerName;
-    public String haMode;  // all or exactly or nodes
-    public String haParams;
-}
