--- driver-rabbitmq/src/main/java/io/openmessaging/driver/rabbitmq/config/RabbitMQClientConfig.java
@@ -1,8 +1,0 @@
-package io.openmessaging.driver.rabbitmq.config;
-
-public class RabbitMQClientConfig {
-    public String user;
-    public String password;
-    public String host;
-    public int port;
-}
