+++ driver-rabbitmq/src/test/java/io/openchaos/driver/rabbitmq/rabbitmq.yaml
@@ -1,0 +1,36 @@
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+name: RabbitMQ
+driverClass: io.openchaos.driver.rabbitmq.RabbitMQDriver
+
+endToEndLatencyCheck: true
+
+# RocketMQ cluster configuration
+
+# Nodes for broker
+nodes:
+  - tcloud # replace with ip or domain name, such as 192.168.0.2
+
+
+# RabbitMQ configuration
+rabbitmqVersion: 3.8.4
+installDir: /usr # you could set existent location for RabbitMQ
+configureFilePath: /rabbitmq/etc
+# RabbitMQ client configuration
+user: root
+password: root
+# RabbitMQ broker configuration
+haMode : classic
+haParms :  {totalNodes : 1 , nodes : 1 }
